+++ rollup.config.js
@@ -3,7 +3,7 @@ import svelte from 'rollup-plugin-svelte';
 import resolve from '@rollup/plugin-node-resolve';
 import livereload from 'rollup-plugin-livereload';
 import { terser } from 'rollup-plugin-terser';
-import css from 'rollup-plugin-css-only';
+//import css from 'rollup-plugin-css-only';
 
 const production = !process.env.ROLLUP_WATCH;
 
@@ -46,12 +46,23 @@ export default {
 		file: 'public/build/bundle.js'
 	},
 	plugins: [
-		svelte({
+		// Two rounds of Svelte - see -> https://github.com/sveltejs/svelte/issues/4228#issuecomment-626315086
+		//
+		svelte({		// web components
+			include: /\/[a-z][^/]+\.svelte$/,
 			compilerOptions: {
-				dev: !production,		// runtime checks
+				dev: !production,
 				customElement: true,
 			}
 		}),
+		svelte({		// normal Svelte classes
+			include: /\/[A-Z][^/]+\.svelte$/,
+			compilerOptions: {
+				dev: !production,
+				customElement: false,
+			}
+		}),
+
 		//css(),
 
 		resolve({